Do not look in remote repositories twice
Earlier, downloading attachments was a special case: it was the only
part of maven-nar-plugin that would look at remote repositories. But
we're now looking into remote repositories by default, so the special
download case looked at the remote repositories twice.

We cannot remove the method NarManager#downloadAttachedNars(List,
List, ArtifactResolver, String), however, because it is part of the
public API of the maven-nar-plugin.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>